requirements.txt
python-dotenv
telethon
requests
flask
